refactor: Controller 추가, main 함수 로직 Controller 로 이동
