+++ src/main/kotlin/lotto/ui/InputView.kt
@@ -1,6 +1,6 @@
 package lotto.ui
 
-internal class InputView {
+class InputView {
     fun inputPurchasePrice(): Int {
         println("구입금액을 입력해 주세요.")
         return readln().toInt()
